requirements.txt
gTTS==1.1.8
python-telegram-bot==5.3.0
